feat: 新增反應身分組功能，包含配置與事件處理
