+++ cfg.yml
@@ -378,3 +378,17 @@ auto_reply:
         Hello World
       match_regex: false          # 是否使用正則表達式匹配
       no_reply: false             # 設為true，則不使用回覆，而是直接傳送訊息
+
+# ================================
+#  ○ 回應身分組
+#  齒輪：reaction_roles.py
+# ================================
+reaction_roles:
+  enabled: true                  # 是否啟用反應身分組
+  role_message_id: 1193435796900954162  # 身分組訊息ID
+  role_channel_id: 1193435796900954162  # 身分組頻道ID
+  reactions:                     # 反應設定，使用清單模式來設定多個
+  - emoji: "🎮"                  # 反應表情符號
+    role_id: 1186541054514704434  # 身分組ID
+  - emoji: "🎨"
+    role_id: 1190290928112517212"
